app/requirements.txt
uvicorn==0.27.0
fastapi==0.109.0
pandas==1.5.3
python-multipart==0.0.6
pymongo==4.4.1
pydantic-settings==2.1.0
requests==2.31.0
geoserver-restconfig==2.0.10
pytest==8.0.0
pytest-cov==4.1.0